+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - \#502 Wrong view loaded if Primary View is selected
 - \#503 Inverted icon colors on 2023.2
 - \#507 Specify id in plugin.xml matching getId() value
+- \#463 Configurable/limit number of displayed builds
 
 ## [0.13.18]
 - \#469 Index out of bounds